2018 Day 06 slight speed up
Replace the use of a map for the grid with a slice.
map access in populateDistances was responsible for ~20% of time.

After this change ~40% of the time is spent on map iteration in
populateDistances for the Locations. This could be turned into a
slice as well, but it is fast enough.